requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Validation
great-expectations==0.18.8

# Model Registry & Serialization
joblib==1.3.2
mlflow==2.8.1

# Evaluation & Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
